Spy handles unresolvable methods as original class
Some framework classes, like UITextField have methods like
isSecureTextEntry which have no implementations and aren't forwarded (at
least on iOS <= 6.1).

This change makes CDRSpy fall back to performing the invocation with the
object as it's original class.  Whenever this happens, messages sent
during the execution of this message won't be recorded.  That's not
ideal but hopefully this will only happen in obscure cases like above.