Add a smoke test for the iOS static lib
We tried to implement this as a separate target inside the existing
Cedar project, but found that XCode 5 has multiple problems adding a
new target for iOS to this project. I suspect it has something to do
with the project having been upgraded from earlier versions of xcode.

It would obviously be best if this was a separate target, but Jesse and
I spent a long time hitting our heads against a wall trying to make that
work, and it seemed like adding this as a separate, standalone project
was the most valuable use of our time on this task.

We tested locally that when a header isn't included in the iOS staticlib
target that the rake task will return a non-zero exit code, so this
should make accepting PRs that run on Travis a lot easier.