[#67321016] developer should not be subject to useless output.
Suppressed NSLog output for Accelerometer warning.

All class initializations are called with pipes are redirected to
/dev/null. This prevents random messages about internals being
logged during the beginning of suites.

We also got rid of the workaround for iOS 6.0 for the
purple workspace port. This doesn't seem to be an issue in 6.1.

Removed CoreLocation framework from spec target.