Make feature specs applescript less fragile
Feature specs now fill out company name field.

These specs should not rely on any of this already being there. It would fail
the first time you ran them on a new machine.

We need to add a new target, but there are no shortcut for this.
Blindly selecting the menu item is prone to failure (because playgrounds were
added and broke these scripts). Using the built in help SEEMS better, although
there are a lot of sleeps in here.