Backout fix for loading test bundle in xcode6
This fails horribly in Xcode5 because the bundler loader is specified twice. It seems that
in Xcode6 the BUNDLE_LOADER shared setting is no longer respected. Their own templates
do not get this, so maybe it was overlooked and will be fixed. For now, feature specs for
templates will pass under Xcode5 but not 6.