OS X, iOS template should show up under correct groups
When we added support for the templates on Xcode 6, somehow the OS X
Spec Suite and the iOS Testing Bundle started showing up under both OS X
and iOS groups for Cedar (which is bad).

Setting the platform for this (which is available in xcode6) seems to
fix it.

Don't install XCTest examples for OS X test bundle.