Fix occasional segfault when deallocating spied objects
This crash was hard to reproduce because it would only occur if the spy
was being deallocated and its memory address was going to be reused
shortly afterwards. In this case, we'd set the isa pointer back to a
CDRSpy, effectively overwriting some memory we no longer own.

Determining if you've been deallocated (or if any object has been, for
that matter) is hard.