Applescript sleep tweaks for slower machines (CI).
More sleeps in the applescript for CI.
Wait for more elements instead of arbitrary sleeps in add_target.applescript.
Wait for Xcode to finish loading the project properly.
Kill Xcode instead of asking it to quit nicely.
Wait for Xcode to launch for slower machines.
Keep trying to open the Xcode project. This fails to open once in CI when
transitioning from the iOS template-project to the OS X project.
Retrying makes it work (could be a timing issue).