rake install no longer depends on ios-sim
Removes the static framework tests from the static framework build
task.  The tests were necessary in older versions of Xcode because
the build process for those targets was unreliable.  Now that Xcode
builds the static framework targets reliably, we no longer need to
run the specs when building them.

We also verified that the static framework target specs still run
as part of the rake ci task.