Update templates to work with Xcode 6.3
Untested with earlier versions of xcode.

Apparently in Xcode 6.3 the TemplateInfo.plist has a version of the macros that
xcode provides (e.g.:: ___FILENAMEBASE___) that are different from the contents of
the template. The template contents use the *actual* file name, rather than what the
user provides.

This fix just captures the class the user typed as a variable, and then uses it
in the template. I believe this should work on earlier versions of Xcode, but
I haven't had the liberty to check yet.

Fixes #324