Create two Cedar.h files. One for OSX, one for iOS.
Rake passes.
There's an extra task here to add a task to the rakefile to test if the
Cedar-iOS-Framework Spec (note the space! This target should be renamed
to avoid confusion with the existing Cedar-iOS-FrameworkSpec! I'm sorry!)
However, I have no idea how to go about modifying the rakefile to check
for that. My apologies.