Don't expose CedarDoubleImpl in Cedar.h
I discovered this while backfilling some specs for BetterConsole earlier
this week. If you check out this commit and try to run the specs, you get
an error message about duplicate symbols for CedarDoubleImpl, which also
matches a class in the Cedar::Doubles C++ namespace.

Pretty sure this is some additional fallout from making Cedar build as a
static framework, but I haven't gone back and tried to bisect the problem.

I have verified that Cedar is still buildable with carthage, and as a cocoapod
and git submodule, so this should be a fairly innocuous change.