Fixing visibility of Cedar headers.
In particular: CDRNil, CedarDoubleImpl, RejectedMethod and StubbedMethod are project headers.

We wanted to move InvocationMatcher.h and NSInvocation+Cedar.h to be project headers as well, to limit
the number of headers exposed to users, but the HaveReceived matcher implementation made this very
difficult.

For reference:

Private headers will be imported by the public headers but are not intended for end users
to import. Project headers will not be copied, and should only be used by implementation (.c, .m, .mm) files.

https://developer.apple.com/library/ios/recipes/xcode_help-project_editor/Articles/Setting%20the%20Role%20of%20a%20Header%20File.html