Fix type utilities spec to use correct-sized pointers
Warning: I'm only about 71.1% sure of what I'm saying here, so
take this with a grain of salt.

I don't think these tests ever passed for the correct reason. The
reason I found this out was because I made some subtle, completely
unrelated changes in the previous commit that increased the size of
the Cedar framework. As a result, a single test here failed, which
caused me to start thinking about pointers to "const char *" and
what it means to be a pointer to NULL.

Previously, we would take our "const char *" argBuffer, dereference
it and ask if the result was NULL. This was find and dandy for most
things, but when we're actually dealing with a "const char **" the
reality is ... more complicated.

The following line is always true so long as the first byte of argBuffer,
when interpreted as a "const char *" is not zero. Given that it points
at a real "const char *" in memory, this seems unlikely, but it's
fairly easy to replicate.

BOOL isNotNull = *argBuffer != '\0';

At first I thought perhaps these pointers were different sizes, but
in fact, both a "const char *" and a "const char **" are 8 bytes on OS X.
Based on this, I believe that the previous code was just checking to see
if the first byte was zero -- which is actually more likely than you
might think!